cloudlinux: task2: almpanis: Data
Completed the second part of the second task of the
internship where I need to create a data
file with read only permissions for normal users
and write only for root and define custom read
and write ops for this file to write only PAGE_SIZE
to it and to return the content on read. A mutex is used
to protect from race conditions.

Signed-off-by: Vasileios Almpanis <vasilisalmpanis@gmail.com>